Added atomic add to ErrorData in compute shader
This meant that the ErrorData array could be cut down to size 1, and
only one GetData call (of an array now at size 1) is necessary because
the direction error difference can be computed directly in the shader.
Eventhough this introduces very expensive synchronization with the
atomic add, it's still faster than getting the large ErrorData array
twice. On the Lenovo laptop it took 39.4 s, now it takes 34.4 seconds.
Some imprecision is introduced because we are converting floats to ints
and back, it's unclear how bad this is, but it does seem to give worse
fitting results.